+++ lib/data/mapper/dashboard/balance_mapper.dart
@@ -1,7 +1,7 @@
 import 'package:zippy/data/api/api_balance.dart';
 
 class BalanceMapper {
-  static double fromApi(ApiBalance apiBalance) {
+  static int fromApi(ApiBalance apiBalance) {
     return apiBalance.balance;
   }
 }
